labs/lab_10_secrets/requirements.txt
python-dotenv==1.0.0
cryptography==41.0.7
pyyaml==6.0.1
python-jose==3.3.0
hvac==1.2.0
pytest==7.4.3 